+++ packages/api/src/modules/applicationFlat/@types/applicationFlatProvider.ts
@@ -1,7 +1,5 @@
-import type { ReadableStream } from "node:stream/web";
-import { ApplicationFlatEntity } from "../../../entities/ApplicationFlatEntity";
+import type { ApplicationFlatEntity } from "../../../entities/ApplicationFlatEntity";
 
 export default interface ApplicationFlatProvider {
-    saveFlatFromStream(stream: ReadableStream<ApplicationFlatEntity>): void;
-    // is supposed to call flatService.saveFromStream
+    saveApplicationsFromStream(stream: ReadableStream<ApplicationFlatEntity>): void;
 }