+++ packages/front/src/lib/helpers/dateHelper.test.js
@@ -8,4 +8,12 @@ describe("DateHelper", () => {
             expect(actual).toEqual(expected);
         });
     });
+
+    describe("withTwoDigitYear", () => {
+        it("return formated year", () => {
+            const expected = "09/12/25";
+            const actual = dateHelper.withTwoDigitYear(new Date("2025/12/09"));
+            expect(actual).toEqual(expected);
+        });
+    });
 });